requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.3.1
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.8
